+++ lib/alias.d.ts
@@ -1,0 +1,3 @@
+// Alias for when you need to use 'any' as a workaround
+// and plan to come back to fix it
+type ShamefulAny = any;