+++ scripts/webpack/webpack.flamegraph.ts
@@ -6,7 +6,7 @@ import { getAlias, getJsLoader, getStyleLoaders } from './shared';
 
 const common = {
   mode: 'production',
-  devtool: 'source-map',
+  //devtool: 'source-map',
 
   resolve: {
     extensions: ['.ts', '.tsx', '.es6', '.js', '.jsx', '.json', '.svg'],
@@ -71,7 +71,7 @@ export default [
     ...common,
     target: 'node',
     mode: 'production',
-    devtool: 'source-map',
+    // devtool: 'source-map',
     entry: {
       index: './src/index.node.ts',
     },
@@ -81,12 +81,17 @@ export default [
       libraryTarget: 'commonjs',
       filename: 'index.node.js',
     },
+
+    externals: {
+      react: 'react',
+      reactDom: 'react-dom',
+    },
   },
   {
     ...common,
     target: 'web',
     mode: 'production',
-    devtool: 'source-map',
+    // devtool: 'source-map',
     entry: {
       index: './src/index.tsx',
     },