+++ webapp/javascript/components/DateRangePicker.tsx
@@ -1,12 +1,17 @@
 import React, { useState } from 'react';
-import { useDispatch } from 'react-redux';
 
+import { useAppDispatch, useAppSelector } from '@pyroscope/redux/hooks';
+import {
+  setDateRange,
+  selectContinuousState,
+} from '@pyroscope/redux/reducers/continuous';
+import cx from 'classnames';
 import Button from '@ui/Button';
 import { faClock } from '@fortawesome/free-solid-svg-icons/faClock';
 import OutsideClickHandler from 'react-outside-click-handler';
 import CustomDatePicker from './CustomDatePicker';
-import { setDateRange } from '../redux/actions';
 import CheckIcon from './CheckIcon';
+import { readableRange } from '../util/formatDate';
 
 const defaultPresets = [
   [
@@ -31,11 +36,27 @@ const defaultPresets = [
   ],
 ];
 
-function DateRangePicker() {
-  const dispatch = useDispatch();
+function findPreset(from: string, until = 'now') {
+  return defaultPresets
+    .flat()
+    .filter((a) => a.until === until)
+    .find((a) => from === a.from);
+}
 
+function dateToLabel(from: string, until: string) {
+  const preset = findPreset(from, until);
+
+  if (preset) {
+    return preset.label;
+  }
+
+  return readableRange(from, until);
+}
+
+function DateRangePicker() {
+  const dispatch = useAppDispatch();
+  const { from, until } = useAppSelector(selectContinuousState);
   const [opened, setOpened] = useState(false);
-  const [range, setRange] = useState();
 
   const toggleDropdown = () => {
     setOpened(!opened);
@@ -44,16 +65,20 @@ function DateRangePicker() {
   const hideDropdown = () => {
     setOpened(false);
   };
-  const selectPreset = ({ from, until }) => {
-    dispatch(setDateRange(from, until));
+  const selectPreset = ({ from, until }: { from: string; until: string }) => {
+    dispatch(setDateRange({ from, until }));
     setOpened(false);
+  };
+
+  const isPresetSelected = (preset: typeof defaultPresets[0][0]) => {
+    return preset.label === dateToLabel(from, until);
   };
 
   return (
     <div className={opened ? 'drp-container opened' : 'drp-container'}>
       <OutsideClickHandler onOutsideClick={hideDropdown}>
         <Button icon={faClock} onClick={toggleDropdown}>
-          {range}
+          {dateToLabel(from, until)}
         </Button>
         <div className="drp-dropdown">
           <div className="drp-quick-presets">
@@ -64,14 +89,15 @@ function DateRangePicker() {
                   {arr.map((x) => (
                     <button
                       type="button"
-                      className={`drp-preset ${
-                        x.label === range ? 'active' : ''
-                      }`}
+                      className={cx(
+                        'drp-preset',
+                        isPresetSelected(x) && 'active'
+                      )}
                       key={x.label}
                       onClick={() => selectPreset(x)}
                     >
                       {x.label}
-                      {x.label === range ? <CheckIcon /> : false}
+                      {isPresetSelected(x) ? <CheckIcon /> : false}
                     </button>
                   ))}
                 </div>
@@ -79,9 +105,11 @@ function DateRangePicker() {
             </div>
           </div>
           <CustomDatePicker
-            setRange={setRange}
-            dispatch={dispatch}
-            setDateRange={setDateRange}
+            from={from}
+            until={until}
+            onSubmit={(from, until) => {
+              dispatch(setDateRange({ from, until }));
+            }}
           />
         </div>
       </OutsideClickHandler>