+++ webapp/javascript/components/Footer.tsx
@@ -1,45 +1,49 @@
 import React from 'react';
-import { connect } from 'react-redux';
 
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faDownload } from '@fortawesome/free-solid-svg-icons/faDownload';
 import { version } from '../../../package.json';
 
-const START_YEAR = 2020;
+const START_YEAR = '2020';
 const PYROSCOPE_VERSION = version;
 
-function copyrightYears(start, end) {
+function copyrightYears(start: string, end: string) {
   return start === end ? start : `${start} – ${end}`;
 }
+
+const win = window as ShamefulAny;
 
 function buildInfo() {
   return `
     BUILD INFO:
     js_version: v${PYROSCOPE_VERSION}
-    goos: ${window.buildInfo.goos}
-    goarch: ${window.buildInfo.goarch}
-    go_version: ${window.buildInfo.goVersion}
-    version: ${window.buildInfo.version}
-    id: ${window.buildInfo.id}
-    time: ${window.buildInfo.time}
-    gitSHA: ${window.buildInfo.gitSHA}
-    gitDirty: ${window.buildInfo.gitDirty}
-    embeddedAssets: ${window.buildInfo.useEmbeddedAssets}
+    goos: ${win.buildInfo.goos}
+    goarch: ${win.buildInfo.goarch}
+    go_version: ${win.buildInfo.goVersion}
+    version: ${win.buildInfo.version}
+    id: ${win.buildInfo.id}
+    time: ${win.buildInfo.time}
+    gitSHA: ${win.buildInfo.gitSHA}
+    gitDirty: ${win.buildInfo.gitDirty}
+    embeddedAssets: ${win.buildInfo.useEmbeddedAssets}
 `.replace(/^\s+/gm, '');
 }
 
 function Footer() {
-  const latestVersion = window.latestVersionInfo.latest_version;
+  const latestVersion = win.latestVersionInfo.latest_version;
   const newVersionAvailable =
-    latestVersion && window.buildInfo.version !== latestVersion;
+    latestVersion && win.buildInfo.version !== latestVersion;
 
   return (
     <div className="footer" title={buildInfo()}>
       <span>
-        {`© Pyroscope ${copyrightYears(START_YEAR, new Date().getFullYear())}`}
+        {`© Pyroscope ${copyrightYears(
+          START_YEAR,
+          new Date().getFullYear().toFixed()
+        )}`}
       </span>
       &nbsp;&nbsp;|&nbsp;&nbsp;
-      <span>{window.buildInfo.version}</span>
+      <span>{win.buildInfo.version}</span>
       {newVersionAvailable && (
         <span>
           &nbsp;&nbsp;|&nbsp;&nbsp;
@@ -58,8 +62,4 @@ function Footer() {
   );
 }
 
-const mapStateToProps = (state) => ({
-  ...state.root,
-});
-
-export default connect(mapStateToProps, {})(Footer);
+export default Footer;