+++ webapp/javascript/components/Settings/Security/ChangePasswordForm.tsx
@@ -8,19 +8,19 @@ import { addNotification } from '@pyroscope/redux/reducers/notifications';
 import StatusMessage from '@ui/StatusMessage';
 import InputField from '@ui/InputField';
 
-function ChangePasswordForm(props) {
+function ChangePasswordForm(props: ShamefulAny) {
   const { user } = props;
-  const [form, setForm] = useState({ errors: [] });
+  const [form, setForm] = useState<ShamefulAny>({ errors: [] });
   const dispatch = useAppDispatch();
   if (user.isExternal) {
     return null;
   }
 
-  const handleChange = (e) => {
+  const handleChange = (e: ShamefulAny) => {
     setForm({ ...form, [e.target.name]: e.target.value });
   };
 
-  const handleFormSubmit = (evt) => {
+  const handleFormSubmit = (evt: ShamefulAny) => {
     evt.preventDefault();
     if (form.password !== form.passwordAgain) {
       return setForm({ errors: ['Passwords must match'] });