+++ webapp/javascript/pages/AdhocComparison.tsx
@@ -7,6 +7,9 @@ import { Tab, Tabs, TabList, TabPanel } from 'react-tabs';
 import Spinner from 'react-svg-spinner';
 import { FlamegraphRenderer } from '@pyroscope/flamegraph';
 import classNames from 'classnames';
+import { Profile } from '@pyroscope/models';
+// eslint-disable-next-line @typescript-eslint/ban-ts-comment
+// @ts-ignore
 import FileList from '../components/FileList';
 import FileUploader from '../components/FileUploader';
 import Footer from '../components/Footer';
@@ -90,7 +93,7 @@ function AdhocComparison() {
                 <FileList
                   className={adhocStyles.tabPanel}
                   profile={leftShared.profile}
-                  setProfile={(p) => {
+                  setProfile={(p: Profile) => {
                     dispatch(setAdhocLeftProfile(p));
                   }}
                 />
@@ -137,7 +140,7 @@ function AdhocComparison() {
                 <FileList
                   className={adhocStyles.tabPanel}
                   profile={rightShared.profile}
-                  setProfile={(p) => {
+                  setProfile={(p: Profile) => {
                     dispatch(setAdhocRightProfile(p));
                   }}
                 />