+++ packages/pyroscope-panel-plugin/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.9](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.2.8...@pyroscope/panel-plugin@1.2.9) (2022-03-15)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
+## [1.2.8](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.2.7...@pyroscope/panel-plugin@1.2.8) (2022-03-14)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
+## [1.2.7](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.2.6...@pyroscope/panel-plugin@1.2.7) (2022-03-14)
+
+**Note:** Version bump only for package @pyroscope/panel-plugin
+
+
+
+
+
 ## [1.2.6](https://github.com/pyroscope-io/pyroscope/compare/@pyroscope/panel-plugin@1.2.5...@pyroscope/panel-plugin@1.2.6) (2022-03-09)
 
 **Note:** Version bump only for package @pyroscope/panel-plugin